Be explicit that we're using Virtuoso. Keep endpoints in VirtuosoRepository
