+++ src/main/resources/application.conf
@@ -280,10 +280,9 @@ cpmeta{
 	}
 
 	fileStoragePath: "./fileStorage"
-	rdfStorage: {
-		sparqlEndpoint: "http://localhost:3030/cpmeta/sparql"
-		updateEndpoint: "http://localhost:3030/cpmeta/sparql"
-		# Username and password are required for write operations to the update endpoint
+
+	virtuoso: {
+		host: "http://localhost:8890"
 		username: dummy
 		password: dummy
 	}
